+++ withnamed.go
@@ -1,0 +1,8 @@
+package imbue
+
+// WithNamedOption is an option that changes the behavior of a call to
+// WithXNamed().
+type WithNamedOption interface {
+	applyWithNamedOptionToContainer(*Container)
+	applyWithNamedOptionToContext(*Context)
+}